Change project-level Gradle to a buildScript
Change the project-level build.gradle.kts to use `buildScript { }` instead of
just a `plugin` block so that we can put and access the Kotlin version in the
AppDependencies.kt file where all of the other versions and dependencies are.
The `plugin` block has to be first and has to use constants and so, prior to
this commit, we were unable to use any values from the AppDependencies.kt file.