Added Room database layer for cache
Add Room database layer to the application for cache. The car crashes are now
returned from the database if there are database crashes saved and retrieved
from the API if there are no database crashes. If the API was used to fetch car
crashes, then the results are saved into the database as a cache for the next
required usage.