Added Google maps UI
Added Google maps to the main screen to be shown. This UI is still a WIP as of
this commit. This work included the reading in of the `MAP_KEY` from
app/app.properties so that the Google Maps API key could be put into the
AndroidManifest file.

The Jetpack Compose Material dependency was also added because it is probable
that, eventually, the app will use the `BottomSheetScaffold` as the "main"
composable. This commit does not make use of this further than:

- Putting the Google Map as the main content

- Putting the old content in the bottom sheet