Add first heatmap to UI
Add a heatmap overlay to the map UI. This required the addition of the Google
Maps Utils dependency. The heatmap is generated from the Latitude and Longitude
of each crash.

The number of crashes requested from the network was increased to 5000 to make
the heatmap show more data.

The crash text UI was set to have a limited max height. This was so that we
could show the HeatMap and not have anything obscured.