Request the latest 3 months of data
This commit rewrites parts of the repository and networking layer to request
the latest 3 months of data instead of a hardcoded amount. When trying to
request more than 3 months of data, there were UI issues as that many LatLng
points were attempted to be loaded into the application.

The repository layer was updated such that the Flow of `carCrashes` is entirely
dependent on the the "hot" exposure of the crash data from the local data
source. The remote data source saves the data into the local data source and
the local data source emits the data as it is saved.

The SimpleDateFormatter was moved to a new "Utils.kt" class since it is now
used in more than one spot and may be used in other spots in the future.