Add visible layout bounds to the map data
Add visible layout bounds to the map data so that only LatLng points that are
within the viewing area of the map can will be reflected in the heatmap and
other ViewModel state data.

Currently this is not working 100% right. Right now the TileOverlay doesn't
have consistent colors when zooming in and out of the map. This is due to
potentially a TileOverlay bug with Jetpack Compose? The TileOverlay isn't
refreshing as it should and isn't respecting the gradient as expected.