Remove Material3 support
Prior to this commit, this project was a mix of Material2 and Material3. The
project was auto-generated with some support for Material3. However, Material2
was also added because that dependency includes some components that were
required. This meant the project was a mix of Material2 and Material3.
Material3 is removed in this project to try to reduce any issues with using
both that might be manifesting themselves in the app.

This commit also includes some light theme redesign to ensure compatibility
with the new Material2 theme.