Refresh data if stale
Add the ability to refresh the data if it is stale. If the data is stale, new
data is requested in which the date range for the requested data is
`NUM_OF_DAYS_BEFORE_LAST_TO_GET` number of days before the most recent car
crash. See the documentation above this variable for more of an explanation
why.

Additionally, `android:backup` was set to `false` in the AndroidManifest.xml to
ensure that the Room database isn't saved across uninstall + reinstall events.

A parameter was added in the repository Flow that ensures that the check to see
if a network request is required is only done the first time.