Error state, repository clean-up
This commit implements an error state with a "retry" button so that the user
can retry getting the data if there's a network error on the initial launch.
This required a change to how the repository works as it was no longer a simple
flow of data from the local data source.

Another change ended up getting pulled in as well: the retrieval of the most
common date of crashes within a given list was changed from a Room database
query to being implemented using repository logic. This decision is entirely
rooted in the author's inability to both stay sane and write complex Room
queries correctly.